requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
pep8==1.7.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
